replace ulong with decimal

+++ 007_hell_or_not/076_relatively_prime_numbers/Program.cs
@@ -2,19 +2,21 @@
 // N ≤ 10²⁰
 
 Console.WriteLine("Задайте натуральное число N (N <= 10^20)");
-// only 1,8...*10^19 (use decimal?)
-ulong limit = ulong.Parse(Console.ReadLine());
+
+decimal limit = decimal.Parse(Console.ReadLine());
 
 Console.WriteLine("Число М групп взаимно простых чисел при разбиении");
 Console.Write("чисел от 1 до N равно ");
-double groups = Math.Floor(Math.Log2(limit) + 1);
+int groups = (int)Math.Floor(Math.Log2((double)limit) + 1);
 Console.WriteLine(groups);
 
 Console.WriteLine("Группы чисел");
 for (int i = 0; i < groups; i++)
 {
     Console.Write($"Группа {i + 1}: ");
-    for (ulong j = Convert.ToUInt64(Math.Pow(2, i)); (j < Convert.ToUInt64(Math.Pow(2, i + 1)) && j <= limit); j++)
+    for (decimal j = Convert.ToDecimal(Math.Pow(2, i)); 
+                (j < Convert.ToDecimal(Math.Pow(2, i + 1)) && j <= limit); 
+                j++)
         Console.Write($"{j} ");
     Console.WriteLine();
 }
